Finished profile page + Start working on instances page in web UI

+++ src/ui/app/routes/instances.py
@@ -15,42 +15,12 @@ instances = Blueprint("instances", __name__)
 @instances.route("/instances", methods=["GET"])
 @login_required
 def instances_page():
-    instances = []
-    instances_types = set()
-    instances_methods = set()
-    instances_healths = set()
-
-    for instance in BW_INSTANCES_UTILS.get_instances():
-        instances.append(
-            {
-                "hostname": instance.hostname,
-                "name": instance.name,
-                "method": instance.method,
-                "health": instance.status,
-                "type": instance.type,
-                "creation_date": instance.creation_date.strftime("%Y-%m-%d at %H:%M:%S %Z"),
-                "last_seen": instance.last_seen.strftime("%Y-%m-%d at %H:%M:%S %Z"),
-            }
-        )
-
-        instances_types.add(instance.type)
-        instances_methods.add(instance.method)
-        instances_healths.add(instance.status)
-
-    # builder = instances_builder(instances, list(instances_types), list(instances_methods), list(instances_healths))
-    # return render_template("instances.html", title="Instances", data_server_builder=b64encode(dumps(builder).encode("utf-8")).decode("ascii"))
-    return render_template("instances.html")  # TODO
+    return render_template("instances.html", instances=BW_INSTANCES_UTILS.get_instances())
 
 
 @instances.route("/instances/new", methods=["PUT"])
 @login_required
 def instances_new():
-    verify_data_in_form(
-        data={"csrf_token": None},
-        err_message="Missing csrf_token parameter on /instances/new.",
-        redirect_url="instances",
-        next=True,
-    )
     verify_data_in_form(
         data={"instance_hostname": None},
         err_message="Missing instance hostname parameter on /instances/new.",
@@ -85,57 +55,41 @@ def instances_new():
     return redirect(url_for("loading", next=url_for("instances.instances_page"), message=f"Creating new instance {instance['hostname']}"))
 
 
-@instances.route("/instances/<string:instance_hostname>", methods=["DELETE"])
+@instances.route("/instances/<string:instance_hostname>/<string:action>", methods=["POST"])
 @login_required
-def instances_delete(instance_hostname: str):
-    verify_data_in_form(
-        data={"csrf_token": None},
-        err_message="Missing csrf_token parameter on /instances/delete.",
-        redirect_url="instances",
-        next=True,
-    )
+def instances_action(instance_hostname: str, action: Literal["ping", "reload", "stop", "delete"]):  # TODO: see if we can support start and restart
+    if action == "delete":
+        delete_instance = None
+        for instance in BW_INSTANCES_UTILS.get_instances():
+            if instance.hostname == instance_hostname:
+                delete_instance = instance
+                break
 
-    delete_instance = None
-    for instance in BW_INSTANCES_UTILS.get_instances():
-        if instance.hostname == instance_hostname:
-            delete_instance = instance
-            break
+        if not delete_instance:
+            return handle_error(f"Instance {instance_hostname} not found.", "instances", True)
+        if delete_instance.method != "ui":
+            return handle_error(f"Instance {instance_hostname} is not a UI instance.", "instances", True)
 
-    if not delete_instance:
-        return handle_error(f"Instance {instance_hostname} not found.", "instances", True)
-    if delete_instance.method != "ui":
-        return handle_error(f"Instance {instance_hostname} is not a UI instance.", "instances", True)
-
-    ret = DB.delete_instance(instance_hostname)
-    if ret:
-        return handle_error(f"Couldn't delete the instance in the database: {ret}", "instances", True)
-
-    return redirect(url_for("loading", next=url_for("instances.instances_page"), message=f"Deleting instance {instance_hostname}"))
-
-
-@instances.route("/instances/<string:action>", methods=["POST"])
-@login_required
-def instances_action(action: Literal["ping", "reload", "stop"]):  # TODO: see if we can support start and restart
-    verify_data_in_form(
-        data={"instance_hostname": None, "csrf_token": None},
-        err_message="Missing instance hostname parameter on /instances/reload.",
-        redirect_url="instances",
-        next=True,
-    )
-
-    DATA["RELOADING"] = True
-    DATA["LAST_RELOAD"] = time()
-    Thread(
-        target=manage_bunkerweb,
-        name=f"Reloading instance {request.form['instance_hostname']}",
-        args=("instances", request.form["instance_hostname"]),
-        kwargs={"operation": action, "threaded": True},
-    ).start()
+        ret = DB.delete_instance(instance_hostname)
+        if ret:
+            return handle_error(f"Couldn't delete the instance in the database: {ret}", "instances", True)
+    else:
+        DATA["RELOADING"] = True
+        DATA["LAST_RELOAD"] = time()
+        Thread(
+            target=manage_bunkerweb,
+            args=("instances", instance_hostname),
+            kwargs={"operation": action, "threaded": True},
+        ).start()
 
     return redirect(
         url_for(
             "loading",
             next=url_for("instances.instances_page"),
-            message=(f"{action.title()}ing" if action != "stop" else "Stopping") + " instance",
+            message=(
+                f"{action.title()}ing"
+                if action not in ("delete", "stop")
+                else ("Deleting" if action == "delete" else "Stopping") + f" instance {instance_hostname}"
+            ),
         )
     )
